switching to Typescript
