+++ .eslintrc.json
@@ -1,0 +1,24 @@
+{
+    "env": {
+        "browser": true,
+        "es6": true,
+        "node": true
+      },
+      "extends": ["eslint:recommended", "prettier"],
+      "plugins": [
+        "prettier"
+      ],
+      "parserOptions": {
+        "ecmaVersion": 2020,
+        "sourceType": "module"
+      },
+      "rules": {
+        "camelcase": [
+          "error"
+        ],
+        "eqeqeq": [
+          "error",
+          "always"
+        ]
+      }
+}
