+++ .prettierrc.json
@@ -1,0 +1,5 @@
+{
+    "tabWidth": 4,
+    "semi": false,
+    "singleQuote": true
+}
