+++ .stylelintrc.yml
@@ -1,0 +1,5 @@
+[
+    'stylelint-config-standard',
+    'stylelint-config-prettier',
+    'stylelint-prettier/recommended',
+]
