+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "cSpell.words": [
+        "camelcase",
+        "eqeqeq",
+        "stylelint",
+        "выиграли",
+        "проиграли"
+    ]
+}
