+++ index.html
@@ -11,15 +11,9 @@
       href="https://fonts.googleapis.com/css2?family=Nunito&display=swap"
       rel="stylesheet"
     />
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="./src/styles/style.css" />
   </head>
   <body>
-    <script src="./index.js"></script>
-    <script src="./generate.js"></script>
-    <script src="./main_page.js"></script>
-    <script src="./popup.js"></script>
-    <script src="./level_1.js"></script>
-    <script src="./level_2.js"></script>
-    <script src="./level_3.js"></script>
+    <script src="dist/bundle.js"></script>
   </body>
 </html>
